// app/types/navigation.ts
export type ViewType =
  | "home"
  | "dashboard"
  | "create"
  | "create-improved"
  | "detail"
  | "mp-demo"
  | "andes-ui-demo"
  | "internal-agreements-dashboard"
  | "agreement-type-selection"
  | "agreement-detail-readonly"
  | "agreement-performance"
  | "audience-management"
  | "approval-status"
  | "add-seller-manual"
  | "upload-sellers-list"
  | "agreement-priorities"
  | "sellers-consultation"
  | "seller-detail"
  | "base-agreement-detail"
  | "seller-agreements-selection"
  | "smb-workflow"
  | "promotions-dashboard"
  | "transactional-query"
  | "promotion-type-selection"
  | "promotion-flow"
  | "seller-mvp-consultation";